Safari-611.1.10/LayoutTests/fast/canvas/draw-focus-if-needed-null-element-expected.txt

Tests that the element parameter to drawFocusIfNeeded() is non-nullable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canvas.drawFocusIfNeeded(null) threw exception TypeError: Argument 1 ('element') to CanvasRenderingContext2D.drawFocusIfNeeded must be an instance of Element.
PASS canvas.drawFocusIfNeeded(path, null) threw exception TypeError: Argument 2 ('element') to CanvasRenderingContext2D.drawFocusIfNeeded must be an instance of Element.
PASS successfullyParsed is true

TEST COMPLETE
